add search bar functionality
