+++ src/components/SearchBar/SearchBar.jsx
@@ -1,9 +1,27 @@
-import React from 'react';
+import React, { useState } from 'react';
 import styles from './SearchBar.module.css';
 
 const SearchBar = () => {
+    const [userList, setUserList] = useState([]);
+    const [userSearch, setUserSearch] = useState('');
+  
+    const handleUserList = () => {
+      setUserList(list => [...list, userSearch]);
+      setUserSearch('');
+    }
+  
+    const handleInputChange = (e) => {
+      setUserSearch(e.target.value);
+    }
     return (
-        <p className={styles.searchBar}>SearchBar</p>
+        <>
+            <input 
+            type="text"
+            value={userSearch}
+            onChange={handleInputChange}
+            />
+            <button onClick={handleUserList}>Buscar</button>
+        </>
     )
 }
